+++ src/Component/Profile/CardUser/CardUser.module.css
@@ -1,0 +1,11 @@
+.img_heder_content {
+    height: 100px;
+    width: 100%;
+    border-radius: 10px;
+  }
+  .user_info {
+    display:flex;
+    justify-content: start;
+    flex-direction: row;
+  
+  }
