+++ src/Component/Profile/MyPost/MyPost.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import mypost from "./MyPost.module.css"
+import Post from "./Post/Post.jsx"
+
+
+const MyPost = ()=> {
+    return (
+        <div>
+        <Post/>
+      </div>
+    )
+}
+export default MyPost
