+++ src/Component/Profile/MyPost/Post/Post.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import post from "./Post.module.css"
+
+const Post = () => {
+    return(
+        <div>
+            <img className={post.img} src="https://pixelbox.ru/wp-content/uploads/2021/05/ava-vk-animal-14.jpg" alt="" />
+        </div>
+    )
+}
+export default Post
