+++ src/Component/Profile/MyPost/Post/Post.module.css
@@ -1,0 +1,4 @@
+.img {
+    width: 50px;
+    border-radius: 10px;
+}
