+++ src/Component/Profile/Profile.module.css
@@ -4,17 +4,7 @@
     background-color: beige;
     border-radius: 10px;
     }
-    .img_heder_content {
-      height: 100px;
-      width: 100%;
-      border-radius: 10px;
-    }
-    .user_info {
-      display:flex;
-      justify-content: start;
-      flex-direction: row;
     
-    }
     
     .user_post  {
       display:flex;
