--- src/Component/css/footer.module.css
@@ -1,5 +1,0 @@
-.footer {
-    grid-area: f;
-    background-image: url(../../img/footer.jpg);
-    height: 140px;
-  }
